--- javascript/nestjs-express/nest/main.d.ts
@@ -1,1 +1,0 @@
-export {};
