--- javascript/nestjs-fastify/app.js
@@ -1,12 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const core_1 = require("@nestjs/core");
-const platform_fastify_1 = require("@nestjs/platform-fastify");
-const app_module_1 = require("./nest/app.module");
-async function bootstrap() {
-    const app = await core_1.NestFactory.create(app_module_1.AppModule, new platform_fastify_1.FastifyAdapter());
-    await app.listen(3000);
-    console.log(`Application is running on: ${await app.getUrl()}`);
-}
-bootstrap();
-//# sourceMappingURL=nest/main.js.map
