+++ javascript/nestjs-fastify/nest-cli.json
@@ -1,0 +1,5 @@
+{
+  "collection": "@nestjs/schematics",
+  "sourceRoot": "src",
+  "entryFile": "app"
+}
