--- javascript/nestjs-fastify/nest/app.controller.d.ts
@@ -1,5 +1,0 @@
-export declare class AppController {
-    getHello(): string;
-    getUserId(id: string): string;
-    postUser(): string;
-}
