--- javascript/nestjs-fastify/nest/app.module.d.ts
@@ -1,2 +1,0 @@
-export declare class AppModule {
-}
